requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
Deprecated==1.2.14
Flask==3.0.1
frozenlist==1.4.1
future==0.18.3
gunicorn==20.0.4
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.7.0
MarkupSafe==2.1.4
multidict==6.0.4
pi==0.1.2
prettyprinter==0.18.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==5.3.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
webexteamssdk==1.6
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
